tests: fix possible 8 minute running time of inotify-rotate.sh
Commit v8.23-63-g111a2b9 removed the expensive tag on this test,
as it runs quickly on systems with inotify.  However without that
it would take about 8 minutes for the test to complete all iterations.

* tests/tail-2/inotify-rotate.sh: Tag as expensive without inotify.
Also adjust the polling parameters used on systems without inotify
so that the test completes within about 15 seconds.