maint: update further copyright year number ranges
* bootstrap: Update copyright year manually (missing in previous
gnulib update).
* tests/init.sh: Likewise.

The entries in the exemption list are processed by
"grep -vEf ./.x-update-copyright", and therefore evaluated as an
extended regular expression (ERE).  Thus, the "bootstrap" entry
also matched for bootstrap.conf which we want to be updated.

* .x-update-copyright: Change all entries to EREs, i.e. including
the caret ^ and dollar sign $ meta-characters matching the beginning
and the end of a line.
* bootstrap.conf: Update copyright year by "make update-copyright".

Finally, the only one showing up with the following command should
be the COPYING file:
  $ git grep 'Copyright .* Free Software' | grep -v '2015 Free Software'